requirements.txt
aioitertools==0.12.0
aiosqlite==0.20.0
alembic==1.14.0
annotated-types==0.7.0
anyio==4.6.2.post1
arize-phoenix==5.7.0
arize-phoenix-evals==0.17.3
arize-phoenix-otel==0.6.1
Authlib==1.3.2
cachetools==5.5.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
click==8.1.7
cryptography==43.0.3
Deprecated==1.2.14
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.4
googleapis-common-protos==1.65.0
graphql-core==3.2.5
greenlet==3.1.1
grpc-interceptor==0.15.4
grpcio==1.67.1
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.4
jiter==0.7.0
joblib==1.4.2
Mako==1.3.6
MarkupSafe==3.0.2
numpy==2.1.3
openai==1.54.3
openinference-instrumentation==0.1.18
openinference-semantic-conventions==0.1.12
opentelemetry-api==1.28.0
opentelemetry-exporter-otlp==1.28.0
opentelemetry-exporter-otlp-proto-common==1.28.0
opentelemetry-exporter-otlp-proto-grpc==1.28.0
opentelemetry-exporter-otlp-proto-http==1.28.0
opentelemetry-proto==1.28.0
opentelemetry-sdk==1.28.0
opentelemetry-semantic-conventions==0.49b0
pandas==2.2.3
protobuf==5.28.3
psutil==6.1.0
pyarrow==18.0.0
pycparser==2.22
pydantic==2.9.2
pydantic_core==2.23.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.17
pytz==2024.2
requests==2.32.3
scikit-learn==1.5.2
scipy==1.14.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.36
sqlean.py==3.47.0
starlette==0.41.2
strawberry-graphql==0.243.1
threadpoolctl==3.5.0
tqdm==4.67.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
uvicorn==0.32.0
websockets==13.1
wrapt==1.16.0
zipp==3.20.2
